requirements.txt
# Core dependencies
requests>=2.28.0
influxdb-client>=1.36.0
prometheus-client>=0.14.1
python-dotenv>=1.0.0